Cleanup "/share" after installation
osinfo-db is not a user facing project, therefore it doesn't ship
appstream data. For osinfo-db to be turned into a Flatpak extension,
it needed the appdata installed in this manifest.

The funny thing is that the Osinfo.Loader object will crawl for XML
files in the osinfo-db source directory and attempt to load them
into the database.

The appstream metainfo file added by this extension was left in the
osinfo-db directory after installation, causing libosinfo to error
out with "Incorrect root element". Which is expected since all osinfo-db
entries should be within the scope of <libosinfo>...</libosinfo>[0].

The "Incorrect root element" error wasn't very informative, but running
"osinfo-db-validate --system" gave it away that this specific file was
incorrect given that its root element was "component", instead of
"libosinfo".

Heh, parsers.

[0] https://gitlab.com/libosinfo/libosinfo/-/blob/master/osinfo/osinfo_loader.c#L1870