manifest: Add flatpak-external-data-checker integration
This will automatically open PRs when new tags appear in the upstream
repository.

This uses the JSON checker rather than the Git checker because the Git
checker cannot [currently][0] extract the timestamp from the Git tag, and
would instead use today's date. fedc uses the Git tree hash to
avoid duplicate PRs, so this would mean a new PR would be opened every
day until one of them were merged.

Unfortunately the JSON provided by GitLab does not include the date from
the tag, only the date from the commit that the tag points to. As we
will see in the subsequent commit, these are different, and it looks a
bit weird that the release date is before the release name (itself a
date). But this is relatively harmless, and importantly it is
deterministic over time.

[0]: https://github.com/flathub/flatpak-external-data-checker/issues/154